refactor: add TypeScript configuration for tests and update some schema types
