+++ api/package.json
@@ -13,6 +13,7 @@
     "#types/*": "./types/*"
   },
   "devDependencies": {
+    "@types/fs-extra": "^11.0.4",
     "@types/memoizee": "^0.4.11",
     "@types/resolve-path": "^1.4.3",
     "@types/semver": "^7.5.8",
