+++ tsconfig.json
@@ -10,5 +10,5 @@
     "checkJs": true,
     "skipLibCheck": true
   },
-  "exclude": ["node_modules", "types", "ui", "dev", "data"]
+  "exclude": ["node_modules", ".type", "ui", "dev", "data", "test-it"]
 }
